chore: move diff into own package
